After 2nd round of CMake refactoring
