+++ CMakeLists.txt
@@ -1,11 +1,11 @@
-cmake_minimum_required(VERSION 3.2)
-project(PolusFilepattern)
+cmake_minimum_required(VERSION 3.22)
+project(filepattern)
 set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
-set (filepattern_VERSION_MAJOR 1)
+set (filepattern_VERSION_MAJOR 2)
 set (filepattern_VERSION_MINOR 0)
-set (filepattern_VERSION_BUILD 3)
+set (filepattern_VERSION_BUILD 0)
 set(filepattern_VERSION "${filepattern_VERSION_MAJOR}.${filepattern_VERSION_MINOR}.${filepattern_VERSION_BUILD}")
 
 if(APPLE)
@@ -29,10 +29,6 @@ if("${CMAKE_CXX_COMPILER_ID}" STREQUAL "GNU")
   endif()
 endif()
 
-# set fp headers
-set(fp_HEADERS
-	src/filepattern/cpp/interface/filepattern.h
-)
 #==== Source files
 set(SOURCE  src/filepattern/cpp/pattern.cpp
 			src/filepattern/cpp/interface/filepattern.cpp
@@ -48,9 +44,40 @@ set(SOURCE  src/filepattern/cpp/pattern.cpp
 			src/filepattern/cpp/util/sort.cpp
 			src/filepattern/cpp/util/vector_parser.cpp
 )
+if (NOT BUILD_PYTHON_LIB) # Not taking the setup.py route, just building libs
+	if (NOT MSVC)
+		if (NOT DEFINED CMAKE_CXX_VISIBILITY_PRESET AND
+			NOT DEFINED CMAKE_VISIBILITY_INLINES_HIDDEN)
+		set(CMAKE_CXX_VISIBILITY_PRESET hidden)
+		set(CMAKE_VISIBILITY_INLINES_HIDDEN YES)
+		endif ()
+	endif()
 
-if(BUILD_CPP_LIB)
-	add_library(filepattern SHARED ${SOURCE})
+
+	if(DEFINED filepattern_SHARED_LIB)
+		set(BUILD_SHARED_LIBS ${filepattern_SHARED_LIB})
+	endif()
+	add_library(filepattern ${SOURCE})
+	add_library(filepattern::filepattern ALIAS filepattern)
+	set_target_properties(filepattern PROPERTIES
+						VERSION ${filepattern_VERSION}
+						SOVERSION ${filepattern_VERSION_MAJOR})
+	target_include_directories(
+		filepattern PUBLIC "$<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>")
+	include(GenerateExportHeader)
+	generate_export_header(filepattern EXPORT_FILE_NAME include/filepattern/filepattern_export.h)
+	target_compile_definitions(
+		filepattern PUBLIC "$<$<NOT:$<BOOL:${BUILD_SHARED_LIBS}>>:FILEPATTERN_STATIC_DEFINE>")
+	target_compile_definitions(filepattern PUBLIC FP_CPP_LIB_EXPORT)
+	target_include_directories(
+		filepattern PUBLIC "$<BUILD_INTERFACE:${CMAKE_CURRENT_BINARY_DIR}/include>")
+	target_link_libraries(filepattern PRIVATE stdc++fs)
+
+	string(COMPARE EQUAL "${CMAKE_SOURCE_DIR}" "${CMAKE_CURRENT_SOURCE_DIR}" is_top_level)
+	option(filepattern_INCLUDE_PACKAGING "Include packaging rules for FilePattern" "${is_top_level}")
+	if (filepattern_INCLUDE_PACKAGING)
+		add_subdirectory(packaging)
+	endif ()
 endif()
 
 
@@ -66,20 +93,10 @@ endif()
 
 set(CMAKE_POSITION_INDEPENDENT_CODE ON)
 
-
-add_executable(	test_fp
-				${SOURCE}
-				tests/cpp/main.cpp
-)
-
 if(CMAKE_CXX_COMPILER_ID STREQUAL "GNU")
   if(BUILD_PYTHON_LIB)
   	target_link_libraries(backend PRIVATE stdc++fs)
   endif()
-  if(BUILD_CPP_LIB)
-	target_link_libraries(filepattern PRIVATE stdc++fs)
-endif()
-  target_link_libraries(test_fp PRIVATE stdc++fs)
 endif()
